Add quick search component
